refactor and support for nodejs modules
