+++ src/example/index.html
@@ -1,6 +1,9 @@
 
 
 
+
+
+../example/index.abell
 <!DOCTYPE html>
 <html lang="en">
 <head>
